fix: change types for Image and Color fields to more appropriate types

+++ src/image/image.rs
@@ -1,11 +1,11 @@
 pub trait Image {
-    fn get_pixel(&self, pixel: (i32, i32)) -> Color;
-    fn set_pixel(&mut self, pixel: (i32, i32), color: Color);
+    fn get_pixel(&self, pixel: (usize, usize)) -> Color;
+    fn set_pixel(&mut self, pixel: (usize, usize), color: Color);
 }
 
 pub struct Color {
-    r: i32,
-    g: i32,
-    b: i32,
-    a: i32,
+    r: u32,
+    g: u32,
+    b: u32,
+    a: u32,
 }
