+++ src/image/rgba32image.rs
@@ -8,11 +8,11 @@ pub struct RGBA32Image {
 }
 
 impl Image for RGBA32Image {
-    fn get_pixel(&self, pixel: (i32, i32)) -> Color {
+    fn get_pixel(&self, pixel: (usize, usize)) -> Color {
         todo!();
     }
 
-    fn set_pixel(&mut self, pixel: (i32, i32), color: Color) {
+    fn set_pixel(&mut self, pixel: (usize, usize), color: Color) {
         todo!();
     }
 }
